requirements-dev.txt
pytest-django
pytest-asyncio
trafaret
